Fix indentation issue
