+++ src/App.css
@@ -63,3 +63,4 @@ button {
 ::placeholder {
   color: white;
   opacity: 1;
+}
